feat: move list api keys by project to project module
