+++ packages/api-gateway/src/models/project.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Transform } from 'class-transformer';
 import { IsNotEmpty } from 'class-validator';
-import { CommonProto } from 'juno-proto';
+import { CommonProto, AuthCommonProto, ApiKeyProto } from 'juno-proto';
 
 export class CreateProjectModel {
   @IsNotEmpty()
@@ -41,4 +41,16 @@ export class LinkUserModel {
 
   @ApiProperty({ description: 'The email of the user' })
   email?: string;
+}
+
+export class GetAllApiKeysResponse {
+  @ApiProperty({
+    isArray: true,
+    description: 'List of API keys belonging to a project',
+  })
+  keys: AuthCommonProto.ApiKey[];
+
+  constructor(res: ApiKeyProto.GetAllApiKeysResponse) {
+    this.keys = res.keys;
+  }
 }
