+++ packages/api-gateway/src/modules/project/project.module.ts
@@ -87,6 +87,7 @@ export class ProjectModule implements NestModule {
         { path: 'project', method: RequestMethod.POST },
         { path: 'project', method: RequestMethod.GET },
         { path: 'project/:id/users', method: RequestMethod.GET },
+        { path: 'project/:id/keys', method: RequestMethod.GET },
         { path: 'project/id/:id', method: RequestMethod.DELETE },
       );
   }
